feature(all): moved around routes
