+++ app.js
@@ -19,6 +19,7 @@ app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 
 app.use('/', require('./routes/index'));
+app.use('/webhooks', require('./routes/webhooks'));
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
